feat(crypto-helper): jwt: move jwt signature to separate module. add warnings if signature key is too short
