+++ src/jwt/jwt.rs
@@ -1,42 +1,7 @@
-use serde_json::Value;
+use super::signature::JwtSignatureAlgorithm;
 
 pub mod editor;
 pub mod viewer;
-
-const JWT_SIGNATURE_ALGORITHMS: [&str; 1] = ["HS256"];
-
-#[derive(Debug, PartialEq, Eq, Clone)]
-pub enum JwtSignatureAlgorithm {
-    Hs256(String),
-    Unsupported,
-}
-
-impl TryFrom<&Value> for JwtSignatureAlgorithm {
-    type Error = String;
-
-    fn try_from(value: &Value) -> Result<Self, Self::Error> {
-        match value {
-            Value::Null => Err("Invalid jwt signature algorithm: null but string extpected".into()),
-            Value::Bool(_) => Err("Invalid jwt signature algorithm: bool but string extpected".into()),
-            Value::Number(_) => Err("Invalid jwt signature algorithm: number but string extpected".into()),
-            Value::String(value) => {
-                if value == JWT_SIGNATURE_ALGORITHMS[0] {
-                    Ok(Self::Hs256(Default::default()))
-                } else {
-                    Ok(Self::Unsupported)
-                }
-            }
-            Value::Array(_) => Err("Invalid jwt signature algorithm: array but string extpected".into()),
-            Value::Object(_) => Err("Invalid jwt signature algorithm: object but string extpected".into()),
-        }
-    }
-}
-
-impl Default for JwtSignatureAlgorithm {
-    fn default() -> Self {
-        Self::Unsupported
-    }
-}
 
 #[derive(Debug, PartialEq, Eq, Default, Clone)]
 pub struct Jwt {
